requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0
scipy>=1.9.0

fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.0
psycopg2-binary>=2.9.0
schedule>=1.2.0
